refactor(character): 删除旧 JsonObject 角色卡模型，统一到 Pydantic
生产代码早已全部切换到 core/data/models/ 的 Pydantic 模型（services.py、command 文件），
旧 JsonObject 体系（base/ 目录 + dnd5e/ 重复副本）仅剩测试在使用，属于死代码。
json_object.py 保留不动，statistics 模块仍依赖它。
