--- src/plugins/DicePP/module/character/base/__init__.py
@@ -1,8 +1,0 @@
-from .health import HPInfo, CHAR_INFO_KEY_HP, CHAR_INFO_KEY_HP_DICE
-from .spell import SpellInfo
-from .ability import AbilityInfo, CHAR_INFO_KEY_ABILITY, CHAR_INFO_KEY_LEVEL, CHAR_INFO_KEY_EXT, CHAR_INFO_KEY_PROF
-from .ability import ability_list, skill_list, saving_list, attack_list, check_item_list, check_item_index_dict, ext_item_list, ext_item_index_dict
-from .money import MoneyInfo
-from .character import DNDCharInfo, gen_template_char
-
-DC_CHAR_DND = "character_dnd"
