+++ src/plugins/DicePP/module/character/dnd5e/__init__.py
@@ -1,9 +1,2 @@
-from ..base import HPInfo, CHAR_INFO_KEY_HP, CHAR_INFO_KEY_HP_DICE
-from ..base import SpellInfo
-from ..base import AbilityInfo, CHAR_INFO_KEY_ABILITY, CHAR_INFO_KEY_LEVEL, CHAR_INFO_KEY_EXT, CHAR_INFO_KEY_PROF
-from ..base import ability_list, skill_list, saving_list, attack_list, check_item_list, check_item_index_dict, ext_item_list, ext_item_index_dict
-from ..base import MoneyInfo
-from ..base import DNDCharInfo, gen_template_char
-
 from .char_command import CharacterDNDCommand, DC_CHAR_DND
 from .hp_command import HPCommand, DC_CHAR_HP
